chore: move attachments.test.ts in same file
- there were two test files
